drain-to-feat for deck-plans — flat → api/components/hooks/types/utils
Mirrors the ADR 2026-05-28 bulletproof-react segmentation already in place
for vehicle-types and vehicles. Pure relocation: no behaviour change.
