+++ src/data/deck-plans/api/fetchDeckPlans.ts
@@ -1,9 +1,9 @@
-import type { AccessToken } from '../../auth/index.ts';
-import { fetchDeckPlansRequest } from '../../graphql/vehicles/queries/fetchDeckPlans.ts';
-import type { DeckPlanContext } from './deckPlanTypes.ts';
-import type { DeckPlan, Name } from '../vehicle-types/types/vehicleTypeTypes.ts';
-import type { Page } from '../../graphql/paginationTypes.ts';
-import { FETCH_ALL_SIZE } from '../../graphql/paginationTypes.ts';
+import type { AccessToken } from '../../../auth/index.ts';
+import { fetchDeckPlansRequest } from '../../../graphql/vehicles/queries/fetchDeckPlans.ts';
+import type { DeckPlanContext } from '../types/deckPlanTypes.ts';
+import type { DeckPlan, Name } from '../../vehicle-types/types/vehicleTypeTypes.ts';
+import type { Page } from '../../../graphql/paginationTypes.ts';
+import { FETCH_ALL_SIZE } from '../../../graphql/paginationTypes.ts';
 
 interface DeckPlanWire {
   netexId: string;
