--- src/data/deck-plans/deckPlanTypes.ts
@@ -1,5 +1,0 @@
-import type { DeckPlan } from '../vehicle-types/types/vehicleTypeTypes';
-
-export type DeckPlanContext = {
-  deckPlans: DeckPlan[];
-};
