+++ src/data/deck-plans/types/deckPlanTypes.ts
@@ -1,0 +1,5 @@
+import type { DeckPlan } from '../../vehicle-types/types/vehicleTypeTypes';
+
+export type DeckPlanContext = {
+  deckPlans: DeckPlan[];
+};
